requirements.txt
dnspython==2.0.0
retry==0.9.2
